+++ src/app/notice/[organization]/page.tsx
@@ -1,5 +1,5 @@
+import ListLayoutComponent from '@/features/contentList/ui/ListLayoutComponent';
 import { noticeItems } from '@/shared/mock/contentList';
-import ListLayoutComponent from '@/widgets/ListLayoutComponent';
 
 const page = () => {
   return <ListLayoutComponent items={noticeItems} title="공지사항" />;
