+++ src/app/proceeding/[organization]/page.tsx
@@ -1,5 +1,5 @@
+import ListLayoutComponent from '@/features/contentList/ui/ListLayoutComponent';
 import { proceedingItems } from '@/shared/mock/contentList';
-import ListLayoutComponent from '@/widgets/ListLayoutComponent';
 
 const page = () => {
   return <ListLayoutComponent items={proceedingItems} title="회의록" />;
