+++ src/app/rule/[organization]/[id]/page.tsx
@@ -1,8 +1,8 @@
 'use client';
 
-import useDetail from '@/shared/hooks/useDetail';
+import useDetail from '@/features/content/model/useDetail';
+import PdfViewerDetailComponent from '@/features/content/ui/PdfViewerDetailComponent';
 import { ruleItems } from '@/shared/mock/contentList';
-import PdfViewerDetailComponent from '@/widgets/PdfViewerDetailComponent';
 
 const page = () => {
   const ruleItem = useDetail(ruleItems);
