+++ src/app/rule/[organization]/page.tsx
@@ -1,5 +1,5 @@
+import ListLayoutComponent from '@/features/contentList/ui/ListLayoutComponent';
 import { ruleItems } from '@/shared/mock/contentList';
-import ListLayoutComponent from '@/widgets/ListLayoutComponent';
 
 const page = () => {
   return <ListLayoutComponent items={ruleItems} title="학칙/회칙" />;
