+++ src/features/content/model/useDetail.ts
