+++ src/features/content/ui/AccordionComponent.tsx
